Improve shared parent class of Group and Step
Move even more shared things to the shared parent class of Group and
Step and rename it to BaseStep.
